App can override client/server targetFrameRate
Prior to this change FishNet sets Application.targetFrameRate, and
overrides any changes the application makes to its value to, for
example, enable the player to lower frame rate to avoid overheating.

This change adds APIs to ClientManager and ServerManager to set a
desired frame rate that overrides the values stored in the prefab.