Update NetworkCollider.cs
Add Interactable Layer configuration to inspector which is now used instead of the game objects layer matrix.
Add Query Trigger Interaction support. Default for triggers is to collide with other triggers.
Add PhysicsScene support for overlaps.
Add _currentlyEntered, a new hashset that contains the currently entered hits. This allows us to trigger OnEnter again if there is a failure to "enter".
OnEnter now returns a bool. True if the Hit collider successfully entered the collider. If false it will attempt to enter again on the next hit. This fixes an issue with CSP movement reconciling immediately after loading into a scene.
Renamed previous to previouslyHit for clarity.